+++ actlog/apps.py
@@ -7,6 +7,3 @@ class ActLogConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "actlog"
     label = "actlog"
-
-    def ready(self) -> None:
-        from actlog import receivers  # noqa: F401